requirements.txt
# Data process
pdfplumber
pdf2img
unstructured # see https://blog.langchain.dev/semi-structured-multi-modal-rag/
unstructured_inference
unstructured_pytesseract
tesseract
pi_heif
qdrant_client

# model & framework
ollama
prompt-template
langchain
langchain_community
langchain_ollama
langchain_huggingface
langchain_experimental
langchain_chroma
langchain_qdrant

langgraph
sentence-transformers
# retriever
faiss-cpu
semantic-chunkers
open-text-embeddings
sentence-transformers

diskcache

